backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-jwt-extended==4.6.0
pymongo==4.6.1
python-dotenv==1.0.0
bcrypt==4.1.2
cloudinary==1.41.0
pydantic==2.5.3
pydantic-settings==2.1.0
firebase-admin==6.3.0
werkzeug==3.0.1
firebase-storage==0.3.0
